- fixed: [Email module - client] fixed an issue where email items could not be properly marked as spam/no spam when working with MySQL in strict mode (closes CN-451)

+++ src/corelib/php/library/Conjoon/Modules/Groupware/Email/Item/Model/Flag.php
@@ -161,7 +161,7 @@ class Conjoon_Modules_Groupware_Email_Item_Model_Flag extends Conjoon_Db_Table {
             return 0;
         }
 
-        $data = array('is_spam' => (bool)$isSpam);
+        $data = array('is_spam' => (int)((bool)$isSpam));
         $adapter = $this->getAdapter();
         return $this->update($data, array(
             $adapter->quoteInto('groupware_email_items_id = ?', $groupwareEmailItemsId, 'INTEGER'),
